pyproject.toml
[project]
name = "kirkify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "scipy>=1.17.0",
    "streamlit>=1.53.0",
]
